+++ packages/manager/src/features/Firewalls/FirewallDetail/Rules/FirewallRuleTable.test.tsx
