+++ packages/manager/src/features/Firewalls/FirewallDetail/Rules/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FirewallRulesLanding';
